LPS-74729 / LPS-73731 SF: calculate visibility in base portlet; do not need visibility in JSPs; a simpler way to get the default tab.
