+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-web/src/main/resources/META-INF/resources/init.jsp
@@ -36,9 +36,6 @@ taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 String defaultTab = (String)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_DEFAULT_TAB);
 
 String tab = ParamUtil.get(request, "tab", defaultTab);
-boolean workflowDefinitionTabVisible = (boolean)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_VISIBILITY_DEFINITION);
-boolean workflowDefinitionLinkTabVisible = (boolean)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_VISIBILITY_DEFINITION_LINK);
-boolean workflowInstanceTabVisible = (boolean)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_VISIBILITY_INSTANCE);
 
 Map<String, DynamicInclude> dynamicIncludes = (Map<String, DynamicInclude>)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_TAB_DYNAMIC_INCLUDES);
 List<String> tabNames = (List<String>)renderRequest.getAttribute(WorkflowWebKeys.WORKFLOW_TAB_NAMES);
